The Big Refactoring of "plain" DelTiC into DelTiC POLYA.
This prepares the way for additional DelTiC-based qdiscs (ZERNO, BOROSHNE, and KHLIB),
which will require subtle changes in how the sojourn time is calculated.  These would
previously have required their own unique copies of deltic_control() with slightly
different behaviour - a maintenance nightmare.

This also means that the new iproute2 module will be able to find the right module to
load and qdisc to interface with.